refactor: convert file names to dash case
