+++ packages/core/utils/src/dml/__tests__/many-to-many.spec.ts
@@ -1,6 +1,6 @@
 import { expectTypeOf } from "expect-type"
 import { TextSchema } from "../schema/text"
-import { ManyToMany } from "../relations/many_to_many"
+import { ManyToMany } from "../relations/many-to-many"
 
 describe("ManyToMany relationship", () => {
   test("define manyToMany relationship", () => {
