+++ packages/core/utils/src/dml/entity-builder.ts
@@ -2,13 +2,13 @@ import { DmlEntity } from "./entity"
 import { TextSchema } from "./schema/text"
 import { EnumSchema } from "./schema/enum"
 import { JSONSchema } from "./schema/json"
-import { HasOne } from "./relations/has_one"
-import { HasMany } from "./relations/has_many"
+import { HasOne } from "./relations/has-one"
+import { HasMany } from "./relations/has-many"
 import { NumberSchema } from "./schema/number"
 import { BooleanSchema } from "./schema/boolean"
-import { BelongsTo } from "./relations/belongs_to"
-import { DateTimeSchema } from "./schema/date_time"
-import { ManyToMany } from "./relations/many_to_many"
+import { BelongsTo } from "./relations/belongs-to"
+import { DateTimeSchema } from "./schema/date-time"
+import { ManyToMany } from "./relations/many-to-many"
 import type { RelationshipOptions, RelationshipType, SchemaType } from "./types"
 
 /**
